+++ src/table/components/virtualized-table/hooks/use-measure-text.ts
